% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{a2m}
\alias{a2m}
\title{Convert an array to an array of dimension 2 or 1}
\usage{
a2m(A, n = length(dim(A)))
}
\arguments{
\item{A}{An array of dimensions (a_1 x ... x a_{dim(A)})}

\item{n}{An integer between 0 and length(dim(A))}
}
\value{
M the matrix of dimension (a_1 x ... x a_n) x (a_{n+1} x ... x a_{dim(A)})
}
\description{
Convert an array to an array of dimension 2 or 1
}
\examples{
A=array(rnorm(prod(2:5)),2:5);M=a2m(A,2);dim(A);dim(M);dim(a2m(A))
}

