% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK.R
\name{add.rg}
\alias{add.rg}
\title{Add a rotation group indicator to all tables of a list when missing.}
\usage{
add.rg(list.tables, id, rg.name)
}
\arguments{
\item{list.tables}{a list of data.frames (order matter)}

\item{id}{a vector of character strings indicating the variable names for the sample unit primary key.}

\item{rg.name}{a character string}
}
\value{
a list of data.frames with a new variable named \code{rg.name}
}
\description{
Add a rotation group indicator to all tables of a list when missing.
}
