% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK.R
\name{add.rg3}
\alias{add.rg3}
\title{Add a rotation group indicator to a table indicating wheter a unit is present in the previous and next samples.}
\usage{
add.rg3(df_1, df0, df1, id, rg.name = "rg")
}
\arguments{
\item{df_1}{a data frame, the previous table}

\item{df0}{a data frame, the current table}

\item{df1}{a data frame, the next table}

\item{id}{a vector of character strings indicating the variable names for the sample unit primary key.}

\item{rg.name}{a character string}
}
\value{
a list of data.frames with a new variable named \code{rg.name}
}
\description{
Add a rotation group indicator to a table indicating wheter a unit is present in the previous and next samples.
}
\details{
creates a variable named \code{rg.name} that takes values
4 for elements present in the current and next tables only,
3 for elements present in the current table only,
2 for elements present in the previous, current and next tables,
1 for elements present in the previous and current tables only.

depends on  dplyr, tidyr
}
\examples{
df <- expand.grid(x= 1:10, y = 1:10)
 df_1 <- df[sample(100,25),]
 df0 <- df[sample(100,25),]
 df1 <- df[sample(100,25),]
 id=c("x","y")
 add.rg3(df_1,df0,df1,c("x","y"))
}
