% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{AKCPS}
\alias{AKCPS}
\title{Gives K coefficient for unemployed used by the Census}
\usage{
AKCPS()
}
\value{
The vector c(a1=ACPSunemployed(),a2=ACPSemployed(),a3=0,k1=KCPSunemployed(),k2=KCPSemployed(),k3=0)
}
\description{
Gives K coefficient for unemployed used by the Census
}

