% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CoeffGM.matrix}
\alias{CoeffGM.matrix}
\title{Compute the Gauss Markov coefficients for Multivariate Blue for arrays}
\usage{
CoeffGM.matrix(Sigma, X, Xplus = MASS::ginv(X))
}
\arguments{
\item{Sigma}{a p x p matrix}

\item{X}{an n x p  matrix}

\item{Xplus:}{a general inverse of X array}
}
\value{
the coefficients matrix \code{W} such that \code{W\times Y} is the best unbiased linear estimator of \code{\beta} where \code{E[Y]=X\times\beta}
}
\description{
Compute the Gauss Markov coefficients for Multivariate Blue for arrays
}
\examples{
A=array(rnorm(prod(2:5)),2:5);M=a2m(A,2);dim(A);dim(M);dim(a2m(A))
}
