% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CoeffGM}
\alias{CoeffGM}
\title{Compute Gauss Markov coefficient for CPS, matrix version}
\usage{
CoeffGM(Sigma, nmonth = dim(Sigma)[[1]])
}
\arguments{
\item{Sigma}{a Variance covariance array}
}
\value{
a matrix.
}
\description{
Compute Gauss Markov coefficient for CPS, matrix version
}
\examples{
CoeffGM(var())
}
