% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CoeffYF}
\alias{CoeffYF}
\title{Compute Yansaneh-Fuller coefficient for CPS, matrix version}
\usage{
CoeffYF(Sigma, nmonth = dim(Sigma)[[1]])
}
\arguments{
\item{Sigma}{a Variance covariance array}
}
\value{
a matrix.
}
\description{
Compute Yansaneh-Fuller coefficient for CPS, matrix version
}
\examples{
CoeffYF(var())
}

