% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{CPS_AK}
\alias{CPS_AK}
\title{Gives A,K coefficient for unemployed used by the Census}
\usage{
CPS_AK()
}
\value{
The vector c(a1=CPS_A_u(),a2=CPS_A_e(),a3=0,k1=CPS_K_u(),k2=CPS_K_e(),k3=0)
}
\description{
Gives A,K coefficient for unemployed used by the Census
}
