% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CPS_X_array}
\alias{CPS_X_array}
\title{Compute Yansaneh-Fuller X matrix for CPS, array version}
\usage{
CPS_X_array(nmonth)
}
\arguments{
\item{nmonth}{an integer}
}
\value{
a matrix.
}
\description{
Compute Yansaneh-Fuller X matrix for CPS, array version
}
\examples{
CPS_X_matrix(10)
}

