% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CPS_X_matrix}
\alias{CPS_X_matrix}
\title{Compute Yansaneh-Fuller X matrix for CPS}
\usage{
CPS_X_matrix(nmonth)
}
\arguments{
\item{nmonth}{an integer}
}
\value{
a matrix.
}
\description{
Compute Yansaneh-Fuller X matrix for CPS
}
\examples{
CPS_X_matrix(10)
}

