% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CPS_Xplus_array}
\alias{CPS_Xplus_array}
\title{Compute the Moore penrose general inverse of a the Yansaneh Fuller X matrix for CPS, array version}
\usage{
CPS_Xplus_array(X)
}
\arguments{
\item{nmonth}{an integer}
}
\value{
a matrix.
}
\description{
Compute the Moore penrose general inverse of a the Yansaneh Fuller X matrix for CPS, array version
}
\examples{
CPS_Xplus_matrix(10)
}

