\name{CRE-package}
\alias{CRE-package}
\alias{CRE}
\docType{package}
\title{Composite Regression Estimation}
\description{

The package allows to compute Composite Regression Estimation as described in FUller and Rao (2001), and AK estimator, as described in Gurney & Dally (1965).}
\details{
\tabular{ll}{
Package: \tab CRE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-15\cr
License: \tab GPL2\cr
}
The most important functions are 
\code{AK} for AK-estimation and \code{MR} for Composite Regression Estimation
}
\author{
Daniel Bonnéry
Maintainer:  <daniel.bonnery@gmail.com>
}
\references{
Fuller and Rao
Gurney and Dally
Singh
}
