% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{empirical.var}
\alias{empirical.var}
\title{Empirical variance of a collection of arrays.}
\usage{
empirical.var(A, MARGIN, n)
}
\arguments{
\item{A}{An array of dimension d_1 x ... d_p}

\item{MARGIN}{a vector of integers}

\item{n}{the array of dimension a_1 x ... x a_n Y[i_1,...,i_n]=sum(W[i_1,...,i_n,,...,] )}
}
\value{

}
\description{
Empirical variance of a collection of arrays.
}
\examples{
empirical.var()
}
