% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorisedf.R
\name{factorisedf}
\alias{factorisedf}
\title{Convert variables to numeric in dataframe.}
\usage{
factorisedf(dfr, list.y)
}
\arguments{
\item{dfr}{A dataframe}

\item{list.y}{character vector containing the names of the variables to be converted.}
}
\value{
a dataframe
}
\description{
Convert variables to numeric in dataframe.
}
\examples{
factorisedf(Orange,names(Orange))
}
