% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{KCPSunemployed}
\alias{KCPSunemployed}
\title{Gives K coefficient for unemployed used by the Census}
\usage{
KCPSunemployed()
}
\value{
.4
}
\description{
Gives K coefficient for unemployed used by the Census
}

