% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{m2a}
\alias{m2a}
\title{Convert  a matrix to an array}
\usage{
m2a(M, a = dim(M), b = numeric(0))
}
\arguments{
\item{M}{A matrix of dimensions (a_1 x ... x a_{dim(A)})}

\item{a}{A vector of integers (numeric(0) is accepted)}
}
\value{
b A vector of integers (numeric(0) is accepted), such that prod(a)*prod(b)=prod(dim(M))
}
\description{
Convert  a matrix to an array
}
\examples{
M<-matrix(1:(prod(2:5)),prod(2:3),prod(4:5));m2a(M,2:3,4:5);identical(m2a(M),M)
}

