% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{var_lin}
\alias{var_lin}
\title{Gives the variance of an array Y that is a linear transformation AX  of an array X from the coefficients of A  and Sigma=Var[X]}
\usage{
var_lin(A, Sigma)
}
\arguments{
\item{Sigma}{An array of dimension b_1 x ... x b_p x b_1 x ... x b_p}

\item{coeff}{An array of dimension a_1 x ... x a_n x b_1 x ... x b_p}
}
\value{
The variance of the AK estimators from the A,K coefficients and the variance covariance matrix .
}
\description{
Gives the variance of an array Y that is a linear transformation AX  of an array X from the coefficients of A  and Sigma=Var[X]
}
\examples{
 a=c(2,4);b=c(3,10,8);A<-array(rnorm(prod(a)*prod(b)),c(a,b));
 dimnames(A)[1:2]<-lapply(a,function(x){letters[1:x]});names(dimnames(A))[1:2]<-c("d1","d2"); 
 Sigma=array(drop(stats::rWishart(1,df=prod(b),diag(prod(b)))),rep(b,2));
 var_lin(A,Sigma)
}

