% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{varAK3}
\alias{varAK3}
\title{Gives the variance of the AK estimators from the A,K coefficients and the variance covariance matrix of the month in sample estimates}
\usage{
varAK3(ak, Sigma)
}
\arguments{
\item{ak}{A set of 3 A, K coefficients, of the form c(a1=.3,a2=.4,a3=0,k1=.4,k2=.7,k3=0).}

\item{Sigma}{An array of dimension 3 x 8 (number of rotation groups) x number of months x 3 x 8 (number of rotation groups) x number of months.}
}
\value{
The variance of the AK estimators from the A,K coefficients and the variance covariance matrix .
}
\description{
Gives the variance of the AK estimators from the A,K coefficients and the variance covariance matrix of the month in sample estimates
}
\examples{
varAK3(ak=c(a1=.3,a2=.4,a3=0,k1=.4,k2=.7,k3=0), Sigma=array(drop(stats::rWishart(1,df=3*10*8,diag(3*10*8))),rep(c(10,8,3),2)))
}

