% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{W.multi.ak}
\alias{W.multi.ak}
\title{general AK weights as a function of a and k parameters.}
\usage{
W.multi.ak(
  months,
  groups,
  S,
  S_1 = S - 1,
  ak,
  eta0 = length(groups)/length(S),
  eta1 = eta0 - 1,
  rescaled = F
)
}
\arguments{
\item{S}{a vector of  integers indicating the indices of the rotation group in the sample}

\item{ak}{a list of 2-dimension vectors}

\item{nmonth}{an integer, indicating number of months}

\item{ngroups}{: number of groups}
}
\value{
an array of AK coefficients  W[m2,m1,mis1] such that Ak estimate for month m2  is sum(W[y2,,])*Y) where Y[m1,mis1] is direct estimate on mis mis1 for emp stat y1 at month m1.
}
\description{
general AK weights as a function of a and k parameters.
}
\examples{
W.multi.ak(months=1:3,groups=1:8,S=c(2:4,6:8),ak=list(c(a=.2,k=.5),c(a=.2,k=.4))) 
}
