% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS.R
\name{WS}
\alias{WS}
\title{Compute weighted sums}
\usage{
WS(list.tables, weight = 1, list.y = NULL, sep = "_n", dimname1 = "m")
}
\arguments{
\item{list.tables}{A list of dataframes}

\item{weight}{either a real number of a character string indicating the name of the weight variable.}

\item{list.y:}{list of variables whose weighted sum needs to be computed. It can be factor or character variables.}
}
\value{
a dataframe.
}
\description{
Compute weighted sums
}
\examples{
WS(plyr::dlply(CRE_data,.variables=~time),"Sampling.weight",c("Hobby","Status","State"));
WS(plyr::dlply(CRE_data,.variables=~time),"Sampling.weight",character(0));
}
