% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSrg.R
\name{WSrg2}
\alias{WSrg2}
\title{Weighted sums by rotation groups}
\usage{
WSrg2(list.tables, weight, y, rg = "hrmis", rescale = F, dimname1 = "m")
}
\arguments{
\item{list.tables}{a named list of data frames}

\item{weight}{a character string  indicating the variable name or a numerical value}

\item{y}{a character strings indicating one study variable}

\item{rg}{a character string indicating the name of the rotation group.}
}
\value{
an array
}
\description{
Weighted sums by rotation groups
}
\examples{
library(dataCPS)
period<-200501:200512
list.tables<-lapply(data(list=paste0("cps",period),package="dataCPS"),get);
names(list.tables)<-period
Y<-WSrg2(list.tables,"pwsswgt",list.y=c("pemlr","pwsswgt"),rg="hrmis")
Y<-WSrg2(list.tables,"pwsswgt",list.y=c("pemlr"),rg="hrmis")
}
