% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{AK_est}
\alias{AK_est}
\title{AK estimation on array of month in sample estimates}
\usage{
AK_est(Y, S = c(2:4, 6:8), a, k)
}
\arguments{
\item{Y}{an array of dimensions}

\item{S}{a vector of integers, subvector of}

\item{a}{a numeric value}

\item{k}{a numeric value}
}
\value{
an array
}
\description{
AK estimation on array of month in sample estimates
}
\examples{

}
