% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{CPS_AK_est}
\alias{CPS_AK_est}
\title{Gives the variance of the AK estimators from the A,K coefficients and the variance covariance matrix of the month in sample estimates}
\usage{
CPS_AK_est(
  mistotals,
  coeff = CPS_AK_coeff.array.fl(dim(mistotals)[1], ak, simplify = FALSE),
  ak = CPS_AK()
)
}
\arguments{
\item{mistotals}{An array of dimension  nmonth x 8 x 3. mistotals[i,j,k] is the month in sample direct estimate for month i, month in sample j rotation group, and variable k.}

\item{coeff}{An array of coefficients W[ak,y2,m2,y1,mis1,m1] such that AK estimate for coefficients ak, month m2 and employment status y2 is sum(W[ak,y2,m2,,,])*Y[,,]) where mistotals[y1,mis1,m1] is direct estimate on mis mis1 for emp stat y1 at month m1.}

\item{ak:}{an ak coefficients vector or a  list of ak coefficients.}
}
\value{
The variance of the AK estimators from the A,K coefficients and the variance covariance matrix .
}
\description{
Gives the variance of the AK estimators from the A,K coefficients and the variance covariance matrix of the month in sample estimates
}
\examples{
varAK3(ak=c(a1=.3,a2=.4,a3=0,k1=.4,k2=.7,k3=0), Sigma=array(drop(stats::rWishart(1,df=3*10*8,diag(3*10*8))),rep(c(10,8,3),2)))
}
