% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YF.R
\name{CPS_Xplus_matrix}
\alias{CPS_Xplus_matrix}
\title{Compute the Moore penrose general inverse of a the X matrix for CPS}
\usage{
CPS_Xplus_matrix(X)
}
\arguments{
\item{nmonth}{an integer, the number of months}

\item{nvar}{an integer, the number of variables}

\item{nrg}{an integer, the number of rotation groups}

\item{alpha}{a coefficient}
}
\value{
a matrix.
}
\description{
Compute the Moore penrose general inverse of a the X matrix for CPS
}
\examples{
CPS_Xplus_matrix(10)
}
