% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double.R
\name{douuble}
\alias{douuble}
\title{Compute weighted sums}
\usage{
douuble(list.tables, w, id, y)
}
\arguments{
\item{list.tables}{A list of dataframes, order matters.}

\item{w}{either  a real number of a character string indicating the name of the weight variable.}

\item{id}{primary key of the tables, used to merge tables together.}

\item{y}{a string indicating the name of a factor variable common to all tables of list.tables.}
}
\value{
a list of three arrays.
}
\description{
Compute weighted sums
}
\examples{
WS(list(cars),"dist","speed")
}

