% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR.R
\name{MR}
\alias{MR}
\title{Regression Composite estimation}
\usage{
MR(list.tables, w, id, list.xMR = NULL, list.x1 = NULL, list.x2 = NULL,
  list.y = NULL, calibmethod = "linear", Alpha = 0.75, theta = 3/4,
  list.dft.x2 = NULL, dft0.xMR = NULL, mu0 = NULL, Singh = TRUE,
  dispweight = FALSE, analyse = FALSE)
}
\arguments{
\item{list.tables}{A list of dataframes}

\item{w}{either a real number of a character string indicating the name of the weight variable.}

\item{id}{an identifier}

\item{list.xMR}{list of variables used to compute proxy composite regression variable}

\item{list.x1}{list of auxiliary variables used in the cablibration, whose calibrated weighted total has to be equal to initially weithed total}

\item{list.x2}{id list of auxiliary variables used in the cablibration, whose calibrated weighted total has to be equal to initially weithed total}

\item{Alpha}{a vector of alpha values. if alpha="01", this will compute MR3}

\item{theta}{}

\item{list.dft.x2}{id list of auxiliary variables used in the cablibration, whose calibrated weighted total has to be equal to initially weithed total}

\item{list.y:}{list of variables whose weighted sum needs to be computed. It can be factor or character variables.}
}
\value{
a dataframe.
}
\description{
Regression Composite estimation
}
\examples{
MR(list.tables<-plyr::dlply(CRE_data,.variables=~time),w="Sampling.weight",list.xMR="Status",id="Identifier",list.y=c("Hobby","Status","State"))$dfEst;
}
