% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK3.R
\name{W.ak}
\alias{W.ak}
\title{general AK weights as a function of a and k parameters.}
\usage{
W.ak(nmonth, ngroups = 8, S = c(2:4, 6:8), a, k)
}
\arguments{
\item{nmonth}{an integer, indicating number of months}

\item{ngroups}{: number of groups}

\item{S}{a vector of  integers indicating the indices of the rotation group in the sample}
}
\value{
an array of AK coefficients  W[m2,m1,mis1] such that Ak estimate for month m2  is sum(W[y2,,])*Y) where Y[m1,mis1] is direct estimate on mis mis1 for emp stat y1 at month m1.
}
\description{
general AK weights as a function of a and k parameters.
}
\examples{

}
